+++ CHANGELOG.md
@@ -38,6 +38,9 @@
 * Fix bug in `Converger.convergeAreaOfInterest(...)` so that if an `AreaOfInterest` is added when an
   existing `Subscription` already exists, the `AreaOfInterest` is marked as `LOADED` rather than getting
   stuck in the `NOT_ASKED` state.
+* Enhance the `replicant.Transport.Context` interface to allow access to underlying request ids received
+  and transmitted. Also support the recording of "sync" request ids so that synchronization attempts can
+  be verified.
 
 ### [v6.02](https://github.com/realityforge/replicant/tree/v6.02) (2018-07-03)
 [Full Changelog](https://github.com/realityforge/replicant/compare/v6.01...v6.02)